i am super bad at coding a regex
