+++ Shared/standard.c
@@ -79,7 +79,7 @@ int isWordCharacter(char ch)
         return 1;
     if ('A' <= ch && ch <= 'Z')
         return 1;
-    return ch == '_';
+    return ch == '_' ? 1 : 0;
 }
 
 int isValidPath(const char* path)
@@ -90,13 +90,15 @@ int isValidPath(const char* path)
         return 0;
     // check first and last /
     if (path[0] != '/' || path[n - 1] != '/')
-        return 1;
+        return 0;
     int ind = 1;
     while (ind < n) {
         // Subpath should not be empty
-        if (!isWordCharacter(ind))
+        if (!isWordCharacter(path[ind]))
             return 0;
-        while (ind + 1 < n && isWordCharacter(++ind));
+        ++ind;
+        while (ind + 1 < n && isWordCharacter(path[ind]))
+            ++ind;
         if (path[ind] != '/')
             return 0;
         ++ind;
